React UI

+++ API/Controllers/WebsiteController.cs
@@ -1,5 +1,7 @@
-﻿using API.Dto;
+﻿using System.Security.Claims;
+using API.Dto;
 using API.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers;
@@ -15,9 +17,20 @@ public class WebsiteController : ControllerBase
         _websiteService = websiteService;
     }
 
+    [Authorize]
     [HttpPost("create/{userId}")]
     public async Task<IActionResult> CreateWebsite(int userId, WebsiteCreateDto websiteCreateDto)
     {
+        
+        var tokenUserId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        // Check if the userId in the route matches the one in the token
+        if (userId != tokenUserId)
+        {
+            return Unauthorized("Unauthorized access. User ID mismatch.");
+        }
+        
+        
         try
         {
             var website = await _websiteService.CreateWebsiteAsync(userId, websiteCreateDto);
@@ -30,15 +43,25 @@ public class WebsiteController : ControllerBase
     }
     
    
+    [Authorize]
     [HttpGet("getWebsites/{userId}")]
     public async Task<IActionResult> GetWebsitesByUserId(int userId)
     {
+        
+        var tokenUserId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        // Check if the userId in the route matches the one in the token
+        if (userId != tokenUserId)
+        {
+            return Unauthorized("Unauthorized access. User ID mismatch.");
+        }
+        
         try
         {
             var websites = await _websiteService.GetWebsitesByUserIdAsync(userId);
             if (!websites.Any())
             {
-                return NotFound("No websites found for the given user.");
+                return Ok(Enumerable.Empty<WebsiteDto>());
             }
             return Ok(websites);
         }
@@ -48,9 +71,19 @@ public class WebsiteController : ControllerBase
         }
     }
 
+    [Authorize]
     [HttpDelete("delete/{userId}/{websiteId}")]
     public async Task<IActionResult> DeleteWebsite(int userId, int websiteId)
     {
+        
+        var tokenUserId = int.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+        // Check if the userId in the route matches the one in the token
+        if (userId != tokenUserId)
+        {
+            return Unauthorized("Unauthorized access. User ID mismatch.");
+        }
+        
         try
         {
             await _websiteService.DeleteWebsiteAsync(userId, websiteId);
